+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Import Widgets_Labels class' methods from Colony Manager Redux.
 -   Import CacheValue(s) classes from Colony Manager Redux.
 -   Utility methods/types for dealing with arrays and ArrayPools.
+-   CustomBackCompatibility utility class for doing custom type replacements on game load.
 
 ## [0.4.0] 2024-08-10